// src/types.ts
export type Network = 'mainnet' | 'testnet'

export interface EpochNetworkConfig {
  packageId:  string
  treasuryId: string
}